Working front-end

+++ frontend/src/components/WorkoutDetails.js
@@ -1,14 +1,23 @@
 import { useState } from 'react';
 import { useWorkoutsContext} from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext';
 import EditWorkoutForm from "./EditWorkoutForm"
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 const WorkoutDetails = ({ workout }) =>{
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
     const [showEditForm, setShowEditForm] = useState(false);
 
+
     const handleClick = async () =>{
+        if (!user){
+          return
+        }
         const response = await fetch('/workouts/' + workout._id,{
-            method: 'DELETE'
+            method: 'DELETE',
+            headers:{
+              'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
         if(response.ok){
